+++ packages/akashic-cli-serve/src/common/targetServiceUtil.ts
@@ -1,0 +1,4 @@
+
+export function isServiceTypeNicoliveLike(targetService: string): boolean {
+	return (/^nicolive.*/.test(targetService) || targetService === "atsumaru:multi");
+}